refactor: implement lazy loading for heavy components in project detail page to improve performance and reduce initial bundle size
