+++ src/app/projects/[slug]/page.tsx
@@ -1,36 +1,25 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect, useCallback, lazy, Suspense } from "react";
 import { useParams, useSearchParams } from "next/navigation";
 import { IProjectResponse } from "@/interfaces/interfaces";
 import Link from "next/link";
-import MidtransScript from "@/components/MidtransScript";
-import DonateModal from "@/components/DonateModal";
-import FundingProgress from "@/components/FundingProgress";
-import ProjectHeader from "@/components/ProjectHeader";
-import RecentDonations from "@/components/RecentDonations";
-import ProjectSidebar from "@/components/ProjectSidebar";
-import SuccessModal from "@/components/SuccessModal";
 import LoadingSpinner from "@/components/LoadingSpinner";
 import ErrorMessage from "@/components/ErrorMessage";
-import {
-  Menu,
-  X,
-  Heart,
-  TrendingUp,
-  Target,
-  AlertTriangle,
-  Zap,
-  ChevronDown,
-  ChevronUp,
-  FileText,
-  ExternalLink,
-  Eye,
-  Video,
-} from "lucide-react";
-import ProjectChat from "@/components/ProjectChat";
-import LivestreamSection from "@/components/LivestreamSection";
 import Navbar from "@/components/Navbar";
+import { Menu, X, Heart, FileText, ExternalLink, Eye } from "lucide-react";
+
+// Lazy load heavy components to reduce initial bundle size
+const MidtransScript = lazy(() => import("@/components/MidtransScript"));
+const DonateModal = lazy(() => import("@/components/DonateModal"));
+const FundingProgress = lazy(() => import("@/components/FundingProgress"));
+const ProjectHeader = lazy(() => import("@/components/ProjectHeader"));
+const RecentDonations = lazy(() => import("@/components/RecentDonations"));
+const ProjectSidebar = lazy(() => import("@/components/ProjectSidebar"));
+const SuccessModal = lazy(() => import("@/components/SuccessModal"));
+const ProjectChat = lazy(() => import("@/components/ProjectChat"));
+const LivestreamSection = lazy(() => import("@/components/LivestreamSection"));
+const ProjectInsights = lazy(() => import("@/components/ProjectInsights"));
 
 interface IApiUserResp {
   username: string;
@@ -68,13 +57,28 @@ interface SwotCard {
   badge: string;
 }
 
+// Loading fallback components
+const ModalLoadingFallback = () => (
+  <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
+    <div className="bg-white p-6 rounded-lg">
+      <LoadingSpinner message="Loading..." />
+    </div>
+  </div>
+);
+
+const ComponentLoadingFallback = ({ message }: { message: string }) => (
+  <div className="bg-gray-50 border border-gray-300 p-6 rounded-lg text-center">
+    <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-black mx-auto mb-4"></div>
+    <p className="text-gray-600 font-normal">{message}</p>
+  </div>
+);
+
 export default function ProjectDetailPage() {
   const params = useParams();
   const searchParams = useSearchParams();
   const [project, setProject] = useState<IProjectResponse | null>(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-  // const [error, setError] = useState('');
   const [showDonateModal, setShowDonateModal] = useState(false);
   const [showSuccessModal, setShowSuccessModal] = useState(false);
   const [donationAmount, setDonationAmount] = useState(0);
@@ -82,10 +86,6 @@ export default function ProjectDetailPage() {
   const [isOwner, setIsOwner] = useState(false);
   const [currentUser, setCurrentUser] = useState<IApiUserResp | null>(null);
   const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [expandedInsights, setExpandedInsights] = useState<
-    Record<string, boolean>
-  >({});
-
   const fetchProject = useCallback(async () => {
     try {
       const response = await fetch(`/api/projects/${params.slug}`);
@@ -160,27 +160,6 @@ export default function ProjectDetailPage() {
     fetchProject();
   };
 
-  const getInsightIcon = (type: string) => {
-    const icons = {
-      strength: <Zap className="w-5 h-5 text-green-600" />,
-      weakness: <AlertTriangle className="w-5 h-5 text-red-600" />,
-      opportunities: <TrendingUp className="w-5 h-5 text-blue-600" />,
-      threat: <Target className="w-5 h-5 text-orange-600" />,
-    };
-    return (
-      icons[type as keyof typeof icons] || (
-        <Heart className="w-5 h-5 text-gray-600" />
-      )
-    );
-  };
-
-  const toggleInsightExpansion = (key: string) => {
-    setExpandedInsights((prev) => ({
-      ...prev,
-      [key]: !prev[key],
-    }));
-  };
-
   if (loading) {
     return <LoadingSpinner message="Loading project..." />;
   }
@@ -199,33 +178,48 @@ export default function ProjectDetailPage() {
     <>
       <Navbar />
       <div className="min-h-screen bg-white">
-        <MidtransScript />
+        {/* Lazy load MidtransScript - only loads when donation modal opens */}
+        {showDonateModal && (
+          <Suspense fallback={<ModalLoadingFallback />}>
+            <MidtransScript />
+          </Suspense>
+        )}
 
-        {/* Modals */}
-        <SuccessModal
-          isOpen={showSuccessModal}
-          onClose={() => setShowSuccessModal(false)}
-          title="Thank You!"
-          message="Your donation of"
-          amount={donationAmount}
-          showFundingUpdate={true}
-        />
+        {/* Modals with lazy loading */}
+        <Suspense fallback={<ModalLoadingFallback />}>
+          <SuccessModal
+            isOpen={showSuccessModal}
+            onClose={() => setShowSuccessModal(false)}
+            title="Thank You!"
+            message="Your donation of"
+            amount={donationAmount}
+            showFundingUpdate={true}
+          />
+        </Suspense>
 
-        <DonateModal
-          isOpen={showDonateModal}
-          onClose={() => setShowDonateModal(false)}
-          projectId={project._id.toString()}
-          isFundingComplete={project.isFundingComplete}
-          onDonationSuccess={handleDonationSuccess}
-        />
+        <Suspense fallback={<ModalLoadingFallback />}>
+          <DonateModal
+            isOpen={showDonateModal}
+            onClose={() => setShowDonateModal(false)}
+            projectId={project._id.toString()}
+            isFundingComplete={project.isFundingComplete}
+            onDonationSuccess={handleDonationSuccess}
+          />
+        </Suspense>
 
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
           {/* Project Header */}
-          <ProjectHeader
-            name={project.name}
-            location={project.location}
-            imageUrl={project.projectImage}
-          />
+          <Suspense
+            fallback={
+              <ComponentLoadingFallback message="Loading project header..." />
+            }
+          >
+            <ProjectHeader
+              name={project.name}
+              location={project.location}
+              imageUrl={project.projectImage}
+            />
+          </Suspense>
 
           {/* Mobile Sidebar Toggle */}
           <div className="lg:hidden mb-6">
@@ -250,13 +244,6 @@ export default function ProjectDetailPage() {
           <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
             {/* Main Content */}
             <div className="lg:col-span-2 space-y-8">
-              {/* Livestream Indicator (for public users) */}
-              {/* {!isOwner && (
-              <LivestreamIndicator 
-                projectSlug={project.slug}
-              />
-            )} */}
-
               {/* Project Description */}
               <div className="bg-white border border-black rounded-lg p-8">
                 <div className="flex items-center gap-3 mb-6">
@@ -278,19 +265,25 @@ export default function ProjectDetailPage() {
               </div>
 
               {/* Funding Progress Component */}
-              <FundingProgress
-                currentFunding={project.currentFunding}
-                fundingGoal={project.fundingGoal}
-                isFundingComplete={project.isFundingComplete}
-                completedAt={project.completedAt}
-                projectSlug={project.slug}
-              />
+              <Suspense
+                fallback={
+                  <ComponentLoadingFallback message="Loading funding progress..." />
+                }
+              >
+                <FundingProgress
+                  currentFunding={project.currentFunding}
+                  fundingGoal={project.fundingGoal}
+                  isFundingComplete={project.isFundingComplete}
+                  completedAt={project.completedAt}
+                  projectSlug={project.slug}
+                />
+              </Suspense>
 
-              {/* Livestream Section */}
+              {/* Livestream Section - Heavy component with Agora SDK */}
               <div className="bg-white border border-black rounded-lg p-8">
                 <div className="flex items-center gap-3 mb-6">
                   <div className="w-10 h-10 bg-red-600 rounded-full flex items-center justify-center">
-                    <Video className="w-5 h-5 text-white" />
+                    <span className="text-white text-lg">📹</span>
                   </div>
                   <div>
                     <h2 className="text-2xl font-medium text-black">
@@ -302,167 +295,56 @@ export default function ProjectDetailPage() {
                   </div>
                 </div>
 
-                <LivestreamSection
-                  isLoggedIn={isLoggedIn}
-                  projectSlug={params.slug as string}
-                  isOwner={isOwner}
-                  userId={currentUser?.userId || ""}
-                  userName={currentUser?.username || ""}
-                />
+                <Suspense
+                  fallback={
+                    <ComponentLoadingFallback message="Loading livestream..." />
+                  }
+                >
+                  <LivestreamSection
+                    isLoggedIn={isLoggedIn}
+                    projectSlug={params.slug as string}
+                    isOwner={isOwner}
+                    userId={currentUser?.userId || ""}
+                    userName={currentUser?.username || ""}
+                  />
+                </Suspense>
               </div>
 
-              {/* Project Chat */}
-              <ProjectChat projectSlug={project.slug} isOwner={isOwner} />
+              {/* Project Chat - Heavy component with Pusher */}
+              <Suspense
+                fallback={
+                  <ComponentLoadingFallback message="Loading chat..." />
+                }
+              >
+                <ProjectChat projectSlug={project.slug} isOwner={isOwner} />
+              </Suspense>
 
               {/* Recent Donations */}
-              <RecentDonations
-                donations={project.donations}
-                projectSlug={project.slug}
-              />
+              <Suspense
+                fallback={
+                  <ComponentLoadingFallback message="Loading donations..." />
+                }
+              >
+                <RecentDonations
+                  donations={project.donations}
+                  projectSlug={project.slug}
+                />
+              </Suspense>
 
-              {/* AI Insights Section - Updated to match flat design */}
+              {/* AI Insights Section */}
               {project.aiInsights &&
                 Object.keys(project.aiInsights).length > 0 && (
-                  <div className="bg-white border border-black rounded-lg p-8">
-                    <div className="flex items-center gap-3 mb-6">
-                      <div className="w-10 h-10 bg-black rounded-full flex items-center justify-center">
-                        <TrendingUp className="w-5 h-5 text-white" />
-                      </div>
-                      <div>
-                        <h2 className="text-2xl font-medium text-black">
-                          AI Project Analysis
-                        </h2>
-                        <p className="text-sm text-gray-600 font-normal">
-                          Strategic insights powered by artificial intelligence
-                        </p>
-                      </div>
-                    </div>
-
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                      {Object.entries(
+                  <Suspense
+                    fallback={
+                      <ComponentLoadingFallback message="Loading AI insights..." />
+                    }
+                  >
+                    <ProjectInsights
+                      aiInsights={
                         project.aiInsights as Record<string, SwotCard>
-                      )
-                        .filter(
-                          ([key]) =>
-                            key === "strength" || key === "opportunities"
-                        )
-                        .map(([key, insight]) => {
-                          const isExpanded = expandedInsights[key] || false;
-
-                          return (
-                            <div
-                              key={key}
-                              className={`border-2 rounded-lg transition-all duration-300 hover:border-gray-800 bg-white ${
-                                key === "strength"
-                                  ? "border-green-500"
-                                  : "border-blue-500"
-                              }`}
-                            >
-                              {/* Main Card Content */}
-                              <div className="p-6">
-                                {/* Header */}
-                                <div className="flex items-start justify-between mb-4">
-                                  <div className="flex items-center gap-3">
-                                    <div
-                                      className={`w-8 h-8 rounded-lg flex items-center justify-center ${
-                                        key === "strength"
-                                          ? "bg-green-500"
-                                          : "bg-blue-500"
-                                      }`}
-                                    >
-                                      {getInsightIcon(key)}
-                                    </div>
-                                    <div>
-                                      <h3 className="text-lg font-medium text-black capitalize">
-                                        {key === "opportunities"
-                                          ? "Opportunities"
-                                          : key
-                                              .replace(/([A-Z])/g, " $1")
-                                              .trim()}
-                                      </h3>
-                                    </div>
-                                  </div>
-                                  <span
-                                    className={`px-3 py-1 text-xs font-normal rounded-full border ${
-                                      key === "strength"
-                                        ? "bg-green-100 text-green-800 border-green-300"
-                                        : "bg-blue-100 text-blue-800 border-blue-300"
-                                    }`}
-                                  >
-                                    {insight.badge}
-                                  </span>
-                                </div>
-
-                                {/* Always Visible Content */}
-                                <div className="space-y-3">
-                                  <h4 className="font-medium text-black leading-snug">
-                                    {insight.title}
-                                  </h4>
-                                  <p className="text-sm text-gray-600 font-normal leading-relaxed">
-                                    {insight.excerpt}
-                                  </p>
-                                </div>
-
-                                {/* Expandable Content */}
-                                <div
-                                  className={`transition-all duration-300 overflow-hidden ${
-                                    isExpanded
-                                      ? "max-h-96 opacity-100 mt-4"
-                                      : "max-h-0 opacity-0"
-                                  }`}
-                                >
-                                  <div className="pt-3 border-t border-gray-200">
-                                    <p className="text-sm text-gray-500 font-normal leading-relaxed">
-                                      {insight.description}
-                                    </p>
-                                  </div>
-                                </div>
-
-                                {/* Toggle Button */}
-                                <button
-                                  onClick={() => toggleInsightExpansion(key)}
-                                  className={`w-full mt-4 py-2 px-4 rounded-full border transition-all duration-200 flex items-center justify-center gap-2 text-sm font-normal ${
-                                    isExpanded
-                                      ? key === "strength"
-                                        ? "bg-green-100 text-green-800 border-green-300 hover:bg-green-200"
-                                        : "bg-blue-100 text-blue-800 border-blue-300 hover:bg-blue-200"
-                                      : "border-black text-gray-600 hover:border-gray-800 hover:bg-gray-50"
-                                  }`}
-                                >
-                                  {isExpanded ? (
-                                    <>
-                                      Show Less
-                                      <ChevronUp className="w-4 h-4" />
-                                    </>
-                                  ) : (
-                                    <>
-                                      Show Details
-                                      <ChevronDown className="w-4 h-4" />
-                                    </>
-                                  )}
-                                </button>
-                              </div>
-                            </div>
-                          );
-                        })}
-                    </div>
-
-                    {/* Additional Info */}
-                    <div className="mt-6 p-4 bg-gray-100 border border-gray-300 rounded-lg">
-                      <div className="flex items-center gap-2 text-sm text-gray-700">
-                        <div className="w-4 h-4 bg-black rounded-full flex items-center justify-center">
-                          <span className="text-white text-xs">✨</span>
-                        </div>
-                        <span className="font-medium">
-                          AI-Powered Analysis:
-                        </span>
-                        <span className="font-normal">
-                          This assessment highlights the project&apos;s
-                          strengths and growth opportunities
-                        </span>
-                      </div>
-                    </div>
-                  </div>
+                      }
+                    />
+                  </Suspense>
                 )}
 
               {/* Project Proposal Section */}
@@ -561,15 +443,21 @@ export default function ProjectDetailPage() {
                 sidebarOpen ? "block" : "hidden lg:block"
               }`}
             >
-              <ProjectSidebar
-                isFundingComplete={project.isFundingComplete}
-                completedAt={project.completedAt}
-                projectStartDate={project.projectStartDate}
-                projectEndDate={project.projectEndDate}
-                fundraisingEndDate={project.fundraisingEndDate}
-                owner={project.owner || undefined}
-                onDonateClick={() => setShowDonateModal(true)}
-              />
+              <Suspense
+                fallback={
+                  <ComponentLoadingFallback message="Loading sidebar..." />
+                }
+              >
+                <ProjectSidebar
+                  isFundingComplete={project.isFundingComplete}
+                  completedAt={project.completedAt}
+                  projectStartDate={project.projectStartDate}
+                  projectEndDate={project.projectEndDate}
+                  fundraisingEndDate={project.fundraisingEndDate}
+                  owner={project.owner || undefined}
+                  onDonateClick={() => setShowDonateModal(true)}
+                />
+              </Suspense>
             </div>
           </div>
         </div>
